+++ mutt/envlist.h
@@ -3,7 +3,7 @@
  * Private copy of the environment variables
  *
  * @authors
- * Copyright (C) 2018 Richard Russon <rich@flatcap.org>
+ * Copyright (C) 2018-2023 Richard Russon <rich@flatcap.org>
  *
  * @copyright
  * This program is free software: you can redistribute it and/or modify it under
@@ -20,15 +20,14 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef MUTT_MUTT_ENVLIST_H
-#define MUTT_MUTT_ENVLIST_H
+#ifndef MUTT_ENVLIST_H
+#define MUTT_ENVLIST_H
 
 #include <stdbool.h>
 
-void   mutt_envlist_free(void);
-char **mutt_envlist_getlist(void);
-void   mutt_envlist_init(char *envp[]);
-bool   mutt_envlist_set(const char *name, const char *value, bool overwrite);
-bool   mutt_envlist_unset(const char *name);
+void   envlist_free (char ***envp);
+char **envlist_init (char  **envp);
+bool   envlist_set  (char ***envp, const char *name, const char *value, bool overwrite);
+bool   envlist_unset(char ***envp, const char *name);
 
-#endif /* MUTT_MUTT_ENVLIST_H */
+#endif /* MUTT_ENVLIST_H */